+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+ - Docstring for `request` package
+   - Docstring for `AirmoreSession` class
+
 ## [0.1.0a9] - 2018-10-17
 ### Changed
  - Travis will upload to Pypi on only tags